Add base activity, lauch activity; refactor login flow
