+++ app/src/main/kotlin/io/imhungry/di/ActivityModule.kt
@@ -2,7 +2,7 @@ package io.imhungry.di
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.imhungry.ui.MainActivity
+import io.imhungry.ui.LoginActivity
 
 @Module
 abstract class ActivityModule {
@@ -16,5 +16,5 @@ abstract class ActivityModule {
     */
 
     @ContributesAndroidInjector
-    abstract fun contributeMainActivityInjector(): MainActivity
+    abstract fun contributeMainActivityInjector(): LoginActivity
 }
