--- app/src/main/kotlin/io/imhungry/ui/LoggedInActivity.kt
@@ -1,2 +1,0 @@
-package io.imhungry.ui
-
